+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# 0.0.3 -- options with debug logging and other features
+
+1. added low level complete logging `new Eth(provider, { debug: false, logger: console, jsonSpace: 0 })`
+2. more tests
+
 # 0.0.2 -- handle eth_getFilterChanges during Block and Pending Tx filter
 
 1. handle getFilterChanges during BlockFilter and PendingTxFilter.
-
 
 # 0.0.1 -- ethjs-query